All classes are in separate files
‘dynamics’ is separated into 4 Generator classes (each in its own
file). Added a template for a Trajectory Analysator.
