+++ GPElib/DynamicsGenerator.py
@@ -1,12 +1,12 @@
 import numpy as np
 from scipy.optimize import minimize
-from sklearn.linear_model import LinearRegression, Lasso
 
 class DynamicsGenerator(object):
 	def __init__(self, **kwargs):
 		self.J = kwargs.get('J', 1.0)
 		# self.beta = kwargs.get('beta', 0.01)
 		self.beta = kwargs.get('beta', 0.01)
+		self.dimensionality = kwargs.get('dimensionality', 1)
 		self.N_wells = kwargs.get('N_wells', 10)
 		self.e_disorder = kwargs.get('disorder', np.zeros(self.N_wells))
 		self.N_part = kwargs.get('N_part_per_well', 100000)
@@ -49,32 +49,8 @@ class DynamicsGenerator(object):
 		else:
 			return i
 
-	def my_unwrap(self, theta):
-		while np.sum(theta.flatten() > np.pi) > 0:
-			theta[theta > np.pi] -= 2.0 * np.pi
-		while np.sum(theta.flatten() < -np.pi) > 0:
-			theta[theta < -np.pi] += 2.0 * np.pi
-		return theta
-
 	def phase_unwrap(self, theta):
 		return theta
-		# if len(theta.shape) > 1:
-		# 	theta0 = theta
-		# 	theta -= np.mean(theta, axis=1)
-		# 	theta = self.my_unwrap(theta)
-		# 	while (np.max(np.sum(np.abs(theta-theta0),axis=1)) > 0.1) or (np.max(np.abs(np.mean(theta,axis=1))) > 0.1):
-		# 		theta0 = theta
-		# 		theta -= np.mean(theta, axis=1)
-		# 		theta = self.my_unwrap(theta)
-		# else:
-		# 	theta0 = theta
-		# 	theta -= np.mean(theta)
-		# 	theta = self.my_unwrap(theta)
-		# 	while (np.sum(np.abs(theta-theta0)) > 0.1) or (np.abs(np.mean(theta)) > 0.1):
-		# 		theta0 = theta
-		# 		theta -= np.mean(theta)
-		# 		theta = self.my_unwrap(theta)
-		# return theta
 
 	def set_init_XY(self, x, y):
 		self.X[0,:] = x
@@ -351,172 +327,3 @@ class DynamicsGenerator(object):
 			return x1, y1, 1
 		else:
 			return x1, y1, 0
-
-class TwoTrajsGenerator(DynamicsGenerator):
-	def __init__(self, **kwargs):
-		DynamicsGenerator.__init__(self, **kwargs)
-		self.RHO1 = np.zeros((self.n_steps, self.N_wells), dtype=self.FloatPrecision)
-		self.THETA1 = np.zeros((self.n_steps, self.N_wells), dtype=self.FloatPrecision)
-		self.X1 = np.zeros((self.n_steps, self.N_wells), dtype=self.FloatPrecision)
-		self.Y1 = np.zeros((self.n_steps, self.N_wells), dtype=self.FloatPrecision)
-		self.energy1 = np.zeros(self.n_steps, dtype=self.FloatPrecision)
-		self.participation_rate1 = np.zeros(self.n_steps, dtype=self.FloatPrecision)
-		self.effective_nonlinearity1 = np.zeros(self.n_steps, dtype=self.FloatPrecision)
-
-		self.number_of_particles1 = np.zeros(self.n_steps, dtype=self.FloatPrecision)
-		self.distance = np.zeros(self.n_steps, dtype=self.FloatPrecision)
-		self.lambdas = []
-		self.lambdas_no_regr = []
-
-	def calc_traj_shift_matrix_cartesian_XY(self, X, Y, X1, Y1):
-		return np.sqrt(np.sum((X - X1) ** 2 + (Y - Y1) ** 2, axis=1))
-
-	def set_init_XY(self, x, y, x1, y1):
-		DynamicsGenerator.set_init_XY(self, x,y)
-		self.X1[0,:] = x1
-		self.Y1[0,:] = y1
-		self.RHO1[0,:], self.THETA1[0,:] = self.from_XY_to_polar(self.X1[0,:], self.Y1[0,:])
-
-	def set_constants_of_motion(self):
-		DynamicsGenerator.set_constants_of_motion(self)
-		self.energy1, self.number_of_particles1, self.angular_momentum1 = self.calc_constants_of_motion(self.RHO1, self.THETA1, self.X1, self.Y1)
-		self.participation_rate1 = np.sum(self.RHO1 ** 4, axis=1) / (np.sum(self.RHO1 ** 2, axis=1) ** 2)
-		self.effective_nonlinearity1 = self.beta * self.participation_rate1 / self.N_wells
-
-class InstabilityGenerator(TwoTrajsGenerator):
-	def __init__(self, **kwargs):
-		TwoTrajsGenerator.__init__(self, **kwargs)
-		self.polarisation = np.zeros(self.n_steps, dtype=self.FloatPrecision)
-		self.polarisation1 = np.zeros(self.n_steps, dtype=self.FloatPrecision)
-		self.perturb_hamiltonian = kwargs.get('perturb_hamiltonian', False)
-		self.error_beta = kwargs.get('error_beta', 0)
-		self.error_J = kwargs.get('error_J', 0)
-		self.error_disorder = kwargs.get('error_disorder', 0)
-
-	def run_dynamics(self):
-		TwoTrajsGenerator.run_dynamics(self)
-		if self.perturb_hamiltonian:
-			x1, y1 = self.X[-1,:], self.Y[-1,:]
-		else:
-			x1,y1 = self.constant_perturbation_XY(self.X[-1,:],self.Y[-1,:])
-		self.set_init_XY(self.X[0,:], self.Y[0,:], x1, y1)
-		self.reverse_hamiltonian(self.error_J, self.error_beta, self.error_disorder)
-		for i in xrange(1, self.n_steps):
-			if (np.min(self.RHO1[i-1,:] ** 2) < self.threshold_XY_to_polar):
-				psi1 = self.rk4_step_exp_XY(np.hstack((self.X1[i-1, :], self.Y1[i-1, :])))
-				self.X1[i, :] = psi1[:self.N_wells]
-				self.Y1[i, :] = psi1[self.N_wells:]
-				self.RHO1[i, :], self.THETA1[i, :] = self.from_XY_to_polar(self.X1[i, :], self.Y1[i, :])
-				self.X1[i, :], self.Y1[i, :] = self.from_polar_to_XY(self.RHO1[i, :], self.THETA1[i, :])
-			else:
-				psi1 = self.rk4_step_exp(np.hstack((self.RHO1[i-1, :], self.THETA1[i-1, :])))
-				self.RHO1[i, :] = psi1[:self.N_wells]
-				self.THETA1[i, :] = psi1[self.N_wells:]
-				self.X1[i, :], self.Y1[i, :] = self.from_polar_to_XY(self.RHO1[i, :], self.THETA1[i, :])
-		self.reverse_hamiltonian(self.error_J, self.error_beta, self.error_disorder)
-		idx = np.arange(self.n_steps)[::-1]
-		self.distance = self.calc_traj_shift_matrix_cartesian_XY(self.X, self.Y, self.X1[idx,:], self.Y1[idx,:])
-		self.set_constants_of_motion()
-		self.calculate_polarisation()
-		if (np.abs(np.max(np.abs(self.energy - self.E_calibr)) / self.E_calibr) > 0.01) or (np.abs(np.max(np.abs(self.energy1 - self.E_calibr)) / self.E_calibr) > 0.01):
-			self.make_exception('Energy is not conserved during the dynamics\n')
-		if (np.abs(np.max(np.abs(self.number_of_particles - self.N_part)) / self.N_part) > 0.01) or (np.abs(np.max(np.abs(self.number_of_particles1 - self.N_part)) / self.N_part) > 0.01):
-			self.make_exception('Number of particles is not conserved during the dynamics\n')
-		self.calculate_lambdas()
-
-	def calculate_polarisation(self):
-		self.polarisation = np.sum(self.X, axis=1)
-		idx = np.arange(self.n_steps)[::-1]
-		self.polarisation1 = np.sum(self.X1[idx,:], axis=1)
-
-	def calculate_lambdas(self):
-		self.lambdas = []
-		self.lambdas_no_regr = []
-		clf = LinearRegression()
-		fr = self.n_steps / 2
-		to = self.n_steps - 1
-		try:
-			clf.fit(self.T[fr:to].reshape(to-fr,1), np.log(self.distance[fr:to] + 1e-15).reshape(to-fr,1))
-			self.lambdas.append(clf.coef_[0][0])
-		except:
-			print 'Bad Lyapunov lambda'
-			self.lambdas.append(0.)
-		self.lambdas_no_regr.append((np.log(self.distance[to] + 1e-15) - np.log(self.distance[fr] + 1e-15)) / (self.T[to] - self.T[fr]))
-
-class LyapunovGenerator(TwoTrajsGenerator):
-	def __init__(self, **kwargs):
-		TwoTrajsGenerator.__init__(self, **kwargs)
-		self.Lyapunov_EPS = kwargs.get('Lyapunov_EPS', 1e-3)
-		self.instability_stops = []
-		self.reset_steps_duration = kwargs.get('reset_steps_duration', 10000)
-		self.distance_check = []
-
-	def reset_perturbation_XY(self, x0, y0, x1, y1):
-		dst = self.calc_traj_shift_XY(x0, y0, x1, y1)
-		if dst < 1e-5:
-			x1 = x0 + (x1 - x0) * 1e-3/dst
-			y1 = y0 + (y1 - y0) * 1e-3/dst
-		dst = self.calc_traj_shift_XY(x0, y0, x1, y1)
-		x1 = x0 + (x1 - x0) * self.PERT_EPS /dst
-		y1 = y0 + (y1 - y0) * self.PERT_EPS /dst
-		return x1, y1
-
-	def run_dynamics(self):
-		x1, y1 = self.constant_perturbation_XY(self.X[0,:], self.Y[0,:])
-		self.set_init_XY(self.X[0,:], self.Y[0,:], x1, y1)
-
-		self.instability_stops = [0]
-		self.distance_check = [self.calc_traj_shift_XY(self.X[0,:], self.Y[0,:], self.X1[0,:], self.Y1[0,:])]
-		for i in xrange(1, self.n_steps):
-			if (np.min(self.RHO[i-1,:] ** 2) < self.threshold_XY_to_polar) or (np.min(self.RHO1[i-1,:] ** 2) < self.threshold_XY_to_polar):
-				psi = self.rk4_step_exp_XY(np.hstack((self.X[i-1, :], self.Y[i-1, :])))
-				psi1 = self.rk4_step_exp_XY(np.hstack((self.X1[i-1, :], self.Y1[i-1, :])))
-				self.X[i, :] = psi[:self.N_wells]
-				self.Y[i, :] = psi[self.N_wells:]
-				self.RHO[i, :], self.THETA[i, :] = self.from_XY_to_polar(self.X[i, :], self.Y[i, :])
-				self.X1[i, :] = psi1[:self.N_wells]
-				self.Y1[i, :] = psi1[self.N_wells:]
-				self.RHO1[i, :], self.THETA1[i, :] = self.from_XY_to_polar(self.X1[i, :], self.Y1[i, :])
-				# self.X1[i, :], self.Y1[i, :] = self.from_polar_to_XY(self.RHO1[i, :], self.THETA1[i, :])
-			else:
-				psi = self.rk4_step_exp(np.hstack((self.RHO[i-1, :], self.THETA[i-1, :])))
-				psi1 = self.rk4_step_exp(np.hstack((self.RHO1[i-1, :], self.THETA1[i-1, :])))
-				self.RHO[i, :] = psi[:self.N_wells]
-				self.THETA[i, :] = psi[self.N_wells:]
-				self.X[i, :], self.Y[i, :] = self.from_polar_to_XY(self.RHO[i, :], self.THETA[i, :])
-				self.RHO1[i, :] = psi1[:self.N_wells]
-				self.THETA1[i, :] = psi1[self.N_wells:]
-				self.X1[i, :], self.Y1[i, :] = self.from_polar_to_XY(self.RHO1[i, :], self.THETA1[i, :])
-			dist = self.calc_traj_shift_XY(self.X[i,:], self.Y[i,:], self.X1[i,:], self.Y1[i,:])
-			self.distance_check.append(dist)
-			# if (dist > self.Lyapunov_EPS) or (i - self.instability_stops[-1] > self.reset_steps_duration):
-			if (i - self.instability_stops[-1] > self.reset_steps_duration):
-				self.X1[i, :], self.Y1[i, :] = self.reset_perturbation_XY(self.X[i, :], self.Y[i, :], self.X1[i, :], self.Y1[i, :])
-				self.RHO1[i, :], self.THETA1[i, :] = self.from_XY_to_polar(self.X1[i, :], self.Y1[i, :])
-				self.instability_stops.append(i)
-		if self.instability_stops[-1] != self.n_steps:
-			self.instability_stops.append(self.n_steps)
-		self.set_constants_of_motion()
-		self.distance = self.calc_traj_shift_matrix_cartesian_XY(self.X, self.Y, self.X1, self.Y1)
-		if (np.abs(np.max(np.abs(self.energy - self.E_calibr)) / self.E_calibr) > 0.01) or (np.abs(np.max(np.abs(self.energy1 - self.E_calibr)) / self.E_calibr) > 0.01):
-			self.make_exception('Energy is not conserved during the dynamics\n')
-		if (np.abs(np.max(np.abs(self.number_of_particles - self.N_part)) / self.N_part) > 0.01) or (np.abs(np.max(np.abs(self.number_of_particles1 - self.N_part)) / self.N_part) > 0.01):
-			self.make_exception('Number of particles is not conserved during the dynamics\n')
-		if (np.max(self.distance_check) > 1.):
-			self.make_exception('Discontinuity during the calculations (distance b/w trajectories > 1)\n')
-		self.calculate_lambdas()
-
-	def calculate_lambdas(self):
-		self.lambdas = []
-		self.lambdas_no_regr = []
-		clf = LinearRegression()
-		for ist in xrange(np.array(self.instability_stops).shape[0]-1):
-			fr = self.instability_stops[ist] + 1
-			to = self.instability_stops[ist+1] - 1
-			try:
-				clf.fit(self.T[fr:to].reshape(to-fr,1), np.log(self.distance[fr:to] + 1e-15).reshape(to-fr,1))
-				self.lambdas.append(clf.coef_[0][0])
-			except:
-				print 'Bad Lyapunov lambda'
-				self.lambdas.append(0.)
-			self.lambdas_no_regr.append((np.log(self.distance[to] + 1e-15) - np.log(self.distance[fr]+ 1e-15)) / (self.T[to] - self.T[fr]))
